requirements.txt
Django==3.2.12
mysqlclient==2.1.0
Scrapy==2.6.1
Unidecode==1.3.3
python-dotenv==0.19.2